kredz-final/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.5.1
python-dotenv==1.0.1
qrcode[pil]==7.4.2
python-multipart==0.0.9
jinja2==3.1.4
httpx==0.27.0
pydantic[email]
